backend/requirements.txt
# Flask==2.3.3
# Flask-CORS==4.0.0
# python-dotenv==1.0.0
# python-dateutil==2.8.2
# openpyxl==3.1.2
# fpdf2==2.7.5



Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
fpdf2==2.7.5
sendgrid==6.12.5